Fix clang build
Summary: Fix clang build

Test Plan: USE_CLANG=1 make check -j64

Subscribers: andrewkr, dhruba

Differential Revision: https://reviews.facebook.net/D59667

+++ table/block.cc
@@ -131,7 +131,7 @@ void BlockIter::Prev() {
     }
     // Loop until end of current entry hits the start of original entry
   } while (NextEntryOffset() < original);
-  prev_entries_idx_ = prev_entries_.size() - 1;
+  prev_entries_idx_ = static_cast<int32_t>(prev_entries_.size()) - 1;
 }
 
 void BlockIter::Seek(const Slice& target) {
